controllers: change resourceManager interface
With the current resourceManager interface there is no way to
communicate the reconcile.Result back to the reconcilePhases as it only
returns the error.

change the interface signature to return the reconcile.Result as well
with the error.

Why?
The change is required for ocs to ocs where we want to communicate the
reconcile.Result to the reconcilePhases.

While making GPRC calls we will get different error codes and based on
those error codes we will decide whether we need to requeue or not if
requeue when to requeue and this can be done with reconcile.Result. With
the existing design, we can return only errors and reconcile will get
requeue. That's why we are changing the interface to return the
reconcile.Result also.

Usecase:
let's say a user will create a storageCluster CR with the wrong
connection string and while making an onboarding request we will get an
error that says invalid connection string. This means there is no point
in re queueing at all as this can not be solved without user interaction.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>