metrics: removed re-export of ClusterResourceQuota by OCS
Under the hood OverprovisionControl is mapped to OpenShift's
ClusterResourceQuota, which is already exported by OpensShift as
Prometheus metrics. No need to re-export the same metrics values and
confuse the end-user.

If in the future there will be a requirement to export more complex
logic (e.g., percentage) than this patch should be reverted.

Signed-off-by: Shachar Sharon <ssharon@redhat.com>