controllers: exit reconcile loop immediately after updating the SC
while onboarding we are not able to update the status of the
storagecluster and we loose the storagecosumer id, there is not a really
easy way to get the consumer id again. Returning exactly after updating
storagecluster annotations will reduce the chances of onboarding to be
failed.

Update will trigger a new reconcile itself so we don't even have to
requeue at all.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>