Limit OCS operator leaderElection & cache to operator namespace
Currently ocs-operator is relying on the default "AllNamespaces" cache
in controller-runtime, which works by syncing all the Kubernetes
resources in the cluster. This initial informer cache sync is so huge
that it causes a sudden massive spike in the memory usage of the
operator. This causes OOMKilled failures for the ocs-operator pods.

So ocs-operator will now set the leaderElectionNamespace & the Manager
cache to the namespace where the operator is deployed. This will
ensure that the initial cache sync will only sync the resources in the
namespace where the operator is deployed.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>