Pass cephFS kernel mount option to rook via ocsoperator configmap
The cephFS kernel mount option is set as a key in the ocsoperator
configmap. The value of the key is passed to rook as a env variable.
The default is "ms_mode=prefer-crc" and if in transit encryption is
enabled, the value is set to "ms_mode=secure". The rook operator
pod is restarted if the value of the key is changed.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>