Always update the template parameters & objects to keep it up-to-date
Trying to update the whole template object doesn't work. So the template
is not reconciled at all. We want the template to be updated to use new
parameters, any new ceph commands, latest rook-ceph image, etc. So we
need to update the parameters & objects separately to keep the template
up-to-date.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>