Create a storageclass with rxbounce option if kubevirt crd is present
We will watch for the virtualmachines.kubevirt.io crd. If the CRD is
present, We will create a special ceph rbd storageclass for the
virtualization environment. This storageclass will have the rxbounce
option. Soto a bounce buffer will be used during I/O operation. This is
mostly targeted to fix issues with windows VMs using ceph rbd. This
storageclass has all the same options as the default ceph rbd
storageclass with this extra rxbounce option.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>