Fix makefile errors
On some systems for example on our CI systems, GOBIN is not in the PATH.
This causes an error "/bin/sh: controller-gen: command not found" to be
reported on running any make target. Also when checking if
controller-gen or kustomize is already installed, it always reports
that it is not installed. So we need to check if the binary is in the
GOBIN directory.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>