Avoid wrong detection of disk type, such as HDD instead of SSD on vSAN
In some situations, such as vSphere VSAN, the drive media type is
detected wrongly, for instance as HDD instead of SSD. For that, we have
to force BlueStore to use settings designed for SSDs. We have 2 options
for this.

1) bluestore_debug_enforce_settings = "ssd"
but it requires restart

OR

2) bluestore_prefer_deferred_size_hdd = 0
should work right away and no new deferred writes will be enqueued

We are choosing the second option as it does not require osd restart.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>