Fix upgrade issue when storagecluster is not present
The ocsinitialization controller just creates the ocs-operator-config cm
(from which rook-ceph-operator pod takes env values for configuration).
The task of keeping it updated is on the storagecluster controller. But
when a storagecluster is not there and an upgrade happens the configmap
is not updated but the new rook operator looks for the Key for
configuration. So the rook-cpeh-operator pod fails & so does the
upgrade.

The solution is to change the behavior of the handling of the configmap
so that when the storagecluster controller is not present
ocsinitialization controller should own the configmap & keep it updated.
But when the storagecluster is present it should do nothing and leave it
to the storagecluster controller.

This commit also includes a change where we no longer compare and update
the individual keys of the configmap. Instead, we just check the whole
data section of the configmap and update it if required.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>