remove use of mixin for prometheusrule generation
PrometheusRules were generated via mixin to allow re-usability and
customizability of the resources programmatically. However, we have
not encountered cases where it has been proven useful. In other words,
we are not using the features that mixin provides. So, this commit
removes mixin from the metrics module. It reduces the maintenance
overhead and eases contribution.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>