Fix cephCluster invalid spec error in replica-1 due to misspelled field
While defining the default for the field it should have been
kubebuilder in all small letters, but earlier it was defined as
kubeBuilder which was causing the error. As the default 1 was not being
set, 0 was being passed as the count to the cephCluster which was
reporting invalid spec error.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>