Add CSI tolerations to rook-ceph-operator-config via controller
Enhance the configuration process by incorporating CSI tolerations
directly into the rook-ceph-operator-config configmap through the
controller. Previously, end users were required to manually edit the
configmap for adding CSI tolerations. With this update, users can now
seamlessly include CSI tolerations in the storage clusters, and the
ocs-operator will automatically propagate these tolerations to the
configmap.

This improvement simplifies the configuration workflow and ensures that
CSI tolerations are seamlessly applied to the Rook Ceph Operator without
the need for manual intervention in the configmap.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>