Add status field to track toleration modifications
Introduce a new status field to track addition of tolerations in the
rook-ceph-operator-config configmap. The added field status indicates
whether tolerations were successfully added to the configmap. This
information is crucial for ensuring safe deletion, as we will only
proceed with deletion if the tolerations were added by our mechanism.
If they were added by user we will not proceed with the deletion.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>