Add CSI_REMOVE_HOLDER_PODS key and its value in the ocs configmap
set the CSI_REMOVE_HOLDER_PODS to true for new users on version 4.16
onwards and set it to false for older/upgraded user to give them time to
migrate manually.

with version 4.17 we will make the CSI_REMOVE_HOLDER_PODS to true for
all customers older and new both.

If the configmap is being created for the first time we are considering
it as a new install.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>