Set values in ocs-operator-config CM for replica-1 in external mode
In external mode the replica-1 feature is not enabled via the
storageCluster CR. So detect the feature is enabled by checking if the
replica-1 storageClass exists. In external mode the failure domain is
not set in the storageCluster CR, so determine the failure domain key
by using the parameter of the storageClass. Also add a watch for
storageClass in the ocsinitialization controller to detect the
storageClass creation. Accordingly set the values in the CM.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>