bundle: add rook as a dependency to allow upgrades from 4.15 to 4.16
In older clusters, upgrading the odf channel from 4.15 to 4.16 tries to
bring the rook-operator, leading to conflicts with ocs-operator 4.15,
which already contains all Rook components in its 4.15 CSV. This
conflict results in upgrade failures.

To address this issue, the Rook dependency is removed from the
odf-operator and added to the ocs-operator. This solution ensures a
smooth upgrade process, as changing the ocs-operator channel will now
bring Rook along with ocs-operator 4.16 without encountering conflicts.

This commit should be reverted in 4.17 as 4.16 will feature different
CSVs. In 4.17, removing Rook as a dependency in ocs and adding it in the
odf should facilitate smooth installations and upgrades.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>