Don't use bluestore-rdr in cephCluster CR
In 4.17 (RHCS 8.0), bluestore-rdr is not a valid
OSD backingstore. This PR removes the method that was
adding bluestore-rdr to a new cephCluster CR.

If the customer is updating to 4.17 from a previous version,
then operator should remove the `bluestore-rdr` store type
from the cephCluster CR. And also add `yes-really-update-store`
so that rook operator can start the OSD migration.

Signed-off-by: sp98 <sapillai@redhat.com>