Update CONTROLLER_GEN_VERSION to v0.15.0 & KUSTOMIZE_VERSION to v4.5.7
Earlier controller-gen versions were having problem with github actions
after updating them to Go 1.22. With the new controller-gen version
the way crds are generated has changed in terms of how multiline
strings are presented in the generated code. This required a change in
the csv-merger tool. This also includes the update of kustomize version
to v4.5.7 and other generated changes.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>