controllers: fix dynamic crd predicate
if a single predicate reject the event then the runtime doesn't queue
the reconcile for that controller and for dynamic crd feature multiple
crds should be checked for event and at the same time presence of a
single matching crd with corresponding predicate functions should
initiate the reconcile.

this commit uses helper function in predicate and creates the required
logical conditions and refactor the crd predicate to directly take
bools based on crd availability.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>