Create CephCluster before external resources to prevent deadlock
When creating the ceph-rbd-rados-namespace StorageClass,
the process waits for the Ceph Block Pool Rados Namespace to be ready,
which in turn depends on the CephCluster being present. To prevent a
deadlock, the CephCluster must be created first. This commit reorders
the resource creation sequence to ensure the CephCluster is available
before proceeding with the external resources.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>