api: update the odf-info api
the ocs-provider-server endpoint will not always be reachable from
the other side. The user might have configured the networking
differently. To address this, we need to make sure that user
can specify the exported endpoint as a annotation to the
storageCluster and we add and pass the info to the odf-info
config map to be used by MCO in case of establishing DR

Signed-off-by: Rewant Soni <resoni@redhat.com>