add isStorageSystem label to the storagecluster
The labels is required to determine if some crd act as a storagesystem
by the odf-operator.

The code to generate label is placed at 2 places, in the kustomization
and csv-merger tool that is because csv-merger tool picks up the crds
from the bases directly instead of running kustomize. I wanted to get
away from the csv-merger tool at some point thats why i have also added
in the kustomize.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>