Set mon_target_pg_per_osd to 400 only during cluster creation
On existing clusters the value of mon_target_pg_per_osd is by default
set to 100. If we set it to 400, there will be a massive increase in the
number of PGs which will cause rebalancing & data movement.

On existing clusters customers will have to set the value of
mon_target_pg_per_osd to 400 manually on the storageCluster CR
if they want to increase the number of PGs.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>