Follow the replicate size of the default CephBlockPool for the .mgr pool
In a cluster if customer has specified the replication size for the
default CephBlockPool for example 2, then the .mgr pool should also
have the same replication size. This will majorly help in prototyping
ODF for 2 node cluster as for all other pools we have ways to
specify the replication size but not for the .mgr pool.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>