Set mon_target_pg_per_osd based on the resource profile
There were concerns of additional memory load on the OSDs with the
increased mon_target_pg_per_osd value of 400. To mitigate this,
we are setting the mon_target_pg_per_osd value based on the resource
profile. By default for the balanced profile, the value is set to 200.
For the lean profile, the value is set to 100. For the performance
profile the value is set to 400.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>