Halve the CPU requests of daemons for IBM Z platform
On IBM Z platform the individual CPU cores are much more powerful than
other architectures. Hence, we halve the CPU requests of daemons
to optimize resource usage.

Note: Unequal CPU limit & request results in the pod not getting
Guaranteed QOS class. But as per the offline discussion the team
has decided to proceed with this change for IBM Z platform.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>