Collect ocs must-gather to artifact directory in CI env for persistence
Currently the must-gather is collected to a local directory which is
lost after the test run. With this change, the must-gather is collected
to the artifact directory in CI env for persistence. Also remove the
collection of ocp must-gather as all the relevant information is already
inside the ocs must-gather.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>