Specify node affinity/tolerations for components in DaemonPlacements map
Required placements for components should be specified in the
DaemonPlacements map for visibility. As adding default node affinity
for all components does not make sense, as that can result in
duplicate node affinity(osd) & node affinity for components(metrics)
that do not support it. Also any tolerations that need to be added
should be specified in the DaemonPlacements map and than used as it
gives visibility at a single place for all kind of placements.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>