Use lean resource profile for storage cluster in e2e tests
Till now, we were setting empty resource lists for each of the daemons.
I recently caught this while testing another PR.
The cluster types that get created during the e2e tests have 3 workers
each with 8 CPU & 16 GB memory. Technically we could change the aws
instance type to have more CPU & memory, but that would increase the
cost of the tests & the time taken to run the tests.
So, we are using the lean resource profile for the storage cluster in
the e2e tests.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>