consumer: create cephclients for each consumer
The cephClients were being created by the primary consumers, with
the suffix of consumer uid. As a result for MDR scenarios, there
were no ceph users created for the secondary client. As mdr follows
a active active approach we need ceph users to be created for both
consumers which have the same permissions.

Signed-off-by: Rewant Soni <resoni@redhat.com>